Send SIGTERM to all processes in a group
Aptible interprets Procfiles with a shell, which means that for most
containers (i.e. unless they were designed to work around this problem),
PID 1 will be a shell that will dutyfully ignore the SIGTERM we send
when attempting to restart the container.

This is problematic, because it reduces our chances of getting a clean
restart by not letting running processes e.g. clean up their PID files.

So, to increase our odds of success, we send SIGTERM to everyone in the
container (like e.g. Heroku does).