Test that Docker doesn't lose ports after TERM
Since we're sending SIGTERM to containers before Docker starts
restarting them, it's possible (some might even say likely..!) that the
container will exit before Docker has a chance to restart it.

This adds a test to check that Docker can properly handle this scenario,
and does not lose the ports exposed by the container.

This also updates our tests to stop using `--rm` (that will break if we
send TERM first), which is consistent with what we do on Aptible.