Workaround Docker's "container destroyed"
When a container exists right around when Docker was attempting to
restart it, it might get into an awkward state that makes it impossible
to restart. Resolving the issue involves restarting the Docker daemon,
which isn't particularly desirable for us.

So, in order to be able to send SIGTERM to all processes in a cgroup,
without making Docker crazy, we'll just wait for the cgroup to exit
ourselves. If it doesn't, then we'll ask Docker to kill it with
prejudice immediately (i.e. `docker kill -t 0`).

This appears to be resolved in 1.11 (and to be most common on 1.9), but
not all our infrastructure is there yet.

See: https://github.com/docker/docker/issues/12738