Wipe by replacing the AUFS mounted directory
This is a little more risky than deleting files, but it's guaranteed to
behave correctly if we manage to restart the container (so it's a lot
more predictable), whereas the "delete the files" approach could leave
deletion markers around (it could also be unpredictably slow, since we
had to copy everything over).

The downside of this approach is that if we're not careful, we may brick
the container we're trying to restart, which is unfortunate but not
catastrophic, since it means we'll just have to run a standard app
restart instead.