Fallback to inline restart when thread start fails
We have seen a case where CC was unable to spawn new threads, but was
still otherwise functioning.

We *may* have succeeded in restarting the offending process in this case
by allocating a little more memory and delivering a SIGTERM (we wouldn't
have been able to fork and run `docker restart` though), so let's
generally avoid requiring a new thread to perform a restart.