Retry docker stop and docker restart
On Docker 1.6, we occasionally seem to hit a race condition in Docker
that prevents us from stopping or restarting a container immediately
after it exits.

This seems to happen because Docker gets a little confused about the
state of the container: Docker thinks it's running, but it's actually
dead.

It's unclear exactly how long it takes for the container to become
usable again, but looking at the Docker logs, there's a good chance that
it's just a few seconds until Docker realizes the container has exited.

So, this patch updates Captain Comeback so that it retries stopping and
restarting the container for a little while.