Extract restart logic into pluggable adapters
This exposes a new `--restart-adapter` option, which can be used to
configure an arbitrary adapter to restart containers. We'll use this to
coordinate restarts via Sweetness.

With a default configuration, there are no functional changes here, with
the exception that Captain Comeback will now deliver a `SIGKILL` if the
container does not exit in time, whereas until now we had relied on
`docker stop ...` to do so (we still run `docker stop`, though).