RestartTestIntegration: fix some racey tests
In RestartTestIntegration, a number of tests spin up a container then
signal it to restart, but they occasionally fail because the container
isn't in the state we expect yet upon booting.

For example, in the tests that use `EXITS_WITH_TERM_1`, where Tini is
supposed to be PID 1, we occasionally signal before Tini has had time to
set up appropriate signal handlers, and the test fails.

This patch fixes this problem by busy-waiting for a certain number of
PIDs in the cgroup, which is a good approximation for "is PID 1 done
doing whatever it needs to do?"