Monitor eventfd for critical memory pressure
This might give us an advance warning that the system is about to run
out of memory, but needs testing.

At a glance, critical is reached about the same time as regular OOM for
RSS-caused evictions. The idea here is to deploy this and monitor when
it's reached in more subtle scenarios (e.g. lots of caches, etc.).

See: https://lwn.net/Articles/544652/