fsync on memory pressure
In our testing, it appears memory pressure on a cgroup can result from
dirty pages that have not been committed to disk yet. In worst cases,
this can cause the cgroup to run out of RAM even with minimal RSS usage.

It also appears that we get the memory pressure notification a few
seconds before the cgroup will run out of RAM (and before e.g. eCryptfs
operations might return errors). So, if we attempt to fsync just as we
get the memory pressure notification, we might be able to avoid a cgroup
OOM altogether.

This patch implements that.